backend/pyproject.toml
[project]
name = "detection-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.32,<1",
    "sqlalchemy[asyncio]>=2.0,<3",
    "asyncpg>=0.30,<1",
    "pydantic-settings>=2.0,<3",
    "arq>=0.26,<1",
    "redis>=5.0,<6",
    "ultralytics>=8.0,<9",
    "opencv-python-headless>=4.9,<5",
    "passlib[bcrypt]>=1.7,<2",
    "python-jose[cryptography]>=3.3,<4",
    "python-multipart>=0.0.9",
    "email-validator>=2.0,<3",
    "alembic>=1.13,<2",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "pytest-bdd>=7,<8",
    "httpx>=0.27,<1",
    "ruff>=0.8,<1",
    "mypy>=1.13,<2",
]

[tool.ruff]
target-version = "py312"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "C901"]

[tool.ruff.lint.mccabe]
max-complexity = 10

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = []

[build-system]
requires = ["setuptools>=75"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
